ActionFactory cleanup
