+++ app/src/components/details/MetricViewPage.tsx
@@ -2,7 +2,6 @@ import React, { Suspense, useEffect, useMemo, useState } from "react";
 import { useDialogContext } from "../layout/dialogs/DialogContext";
 import { useMetricContext } from "./MetricDetailsContext";
 import { GroupByTime } from "./chart/consolidation/GroupByTime";
-import { IIconButtonAction } from "../common/actions/IconButtonWrapper";
 import { LocalHotelOutlined } from "@mui/icons-material";
 import { getCommonActions } from "../overview/getCommonActions";
 import { PageSection } from "../layout/pages/PageSection";
@@ -22,6 +21,7 @@ import { getDefaultDateConditions } from "./filters/DateFilters";
 import { GenericEmptyPlaceholder } from "../common/search/GenericEmptyPlaceholder";
 import { MyChartType } from "./chart/grouping/ChartTypeSelector";
 import { ActionFactory } from "../common/actions/ActionFactory";
+import { IIconButtonAction } from "../common/actions/IIconButtonAction";
 
 export const MetricViewPage: React.FC = () => {
   const { renderDialog } = useDialogContext();
