+++ app/src/components/overview/MetricHeaderActions.tsx
@@ -2,8 +2,8 @@ import React, { useMemo } from "react";
 import { IMetric } from "../../serverApi/IMetric";
 import { useDialogContext } from "../layout/dialogs/DialogContext";
 import { Actions } from "../common/actions/Actions";
-import { IIconButtonAction } from "../common/actions/IconButtonWrapper";
 import { getCommonActions } from "./getCommonActions";
+import { IIconButtonAction } from "../common/actions/IIconButtonAction";
 
 export const MetricHeaderActions: React.FC<{ metric: IMetric }> = ({
   metric,
